# openapi/proofmark-marketplace.yaml
openapi: "3.0.0"
info:
  title: ProofMark API
  version: "1.0.0"
  description: |
    Proof of creation, sealed forever. ProofMark creates timestamped,
    cryptographically signed evidence of authorship for screenplays,
    manuscripts, and creative works. Protect drafts, generate evidence
    packets, verify hashes, and share work with tracked links.

    ## Authentication
    All authenticated endpoints require a ProofMark API key passed
    via the `X-ProofMark-Key` header. Payment is handled automatically
    via x402 protocol by the Mirra platform.

    ## Mirra Context Headers
    The platform injects `x-mirra-user-id` and `x-mirra-session-id`
    headers on every call. ProofMark uses these to map Mirra users
    to ProofMark accounts (auto-provisioned on first call).
  contact:
    name: ProofMark
    url: https://proofmark.xyz
  license:
    name: Proprietary
servers:
  - url: https://api.proofmark.xyz/v1
    description: Production
paths:
...
  /scripts:
    get:
      operationId: listScripts
      summary: List protected works
      description: |
        Returns all scripts protected by the calling user, sorted by
        most recently updated. Includes latest version metadata for
        each script.
      tags:
        - Scripts
      parameters:
        - name: workType
          in: query
          description: Filter by work type
          schema:
            $ref: "#/components/schemas/WorkType"
        - name: limit
          in: query
          description: Max results (default 50, max 100)
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of protected scripts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptList"
        "402":
          description: Payment required (x402)

  # ─────────────────────────────────────────────
  #  GET SCRIPT — single script details
  # ─────────────────────────────────────────────
...
  /scripts/{scriptId}/packet:
    get:
      operationId: getEvidencePacket
      summary: Generate evidence packet
      description: |
        Returns a ZIP archive containing the complete evidence packet
        for a committed version: original file, SHA-256 receipt,
        metadata summary, USCO filing worksheet, audit log, and
        ProofMark certificate PDF.
      tags:
        - Evidence
      parameters:
        - $ref: "#/components/parameters/scriptId"
        - name: versionId
          in: query
          description: Specific version ID. Defaults to latest committed version.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Evidence packet ZIP
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "400":
          description: Version not committed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "402":
          description: Payment required (x402)
        "404":
          description: Script or version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ─────────────────────────────────────────────
  #  CERTIFICATE — standalone PDF certificate
  # ─────────────────────────────────────────────
  /scripts/{scriptId}/certificate:
    get:
      operationId: getCertificate
      summary: Generate protection certificate
      description: |
        Returns a standalone PDF certificate for a committed version.
        Contains script title, SHA-256 fingerprint, timestamp, and
        blockchain transaction proof (if registered).
      tags:
        - Evidence
      parameters:
        - $ref: "#/components/parameters/scriptId"
        - name: versionId
          in: query
          description: Specific version ID. Defaults to latest committed version.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Certificate PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "402":
          description: Payment required (x402)
        "404":
          description: Script or version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ─────────────────────────────────────────────
  #  SHARE — create tracked share link
  # ─────────────────────────────────────────────
...
  /scripts/{scriptId}/audit:
    get:
      operationId: getAuditLog
      summary: Get audit log
      description: |
        Returns the append-only audit trail for a script. Every
        action — upload, commit, share, view, packet generation —
        is logged with timestamps and metadata.
      tags:
        - Scripts
      parameters:
        - $ref: "#/components/parameters/scriptId"
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLog"
        "402":
          description: Payment required (x402)
        "404":
          description: Script not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

# ═══════════════════════════════════════════════
#  COMPONENTS
# ═══════════════════════════════════════════════
components:
...
  parameters:
    scriptId:
      name: scriptId
      in: path
      required: true
      description: UUID of the protected script
      schema:
        type: string
        format: uuid

  # ── Schemas ─────────────────────────────────
  schemas:
...
    WorkType:
      type: string
      enum:
        - screenplay
        - manuscript
        - treatment
        - pitch_deck
        - research
        - other
      description: The type of creative work
    ChainStatus:
      type: string
      enum:
        - pending
        - confirmed
        - failed
        - null
      description: Blockchain registration status
...
    ScriptSummary:
      type: object
      properties:
        scriptId:
          type: string
          format: uuid
        title:
          type: string
        workType:
          $ref: "#/components/schemas/WorkType"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        latestVersion:
          type: object
          properties:
            versionId:
              type: string
              format: uuid
            versionNumber:
              type: integer
            sha256:
              type: string
            committedAt:
              type: string
              format: date-time
            chainStatus:
              $ref: "#/components/schemas/ChainStatus"
            byteSize:
              type: integer
            pageCount:
              type: integer
              nullable: true
    ScriptList:
      type: object
      properties:
        scripts:
          type: array
          items:
            $ref: "#/components/schemas/ScriptSummary"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    AuditLogEntry:
      type: object
      properties:
        action:
          type: string
          description: Action type
          enum:
            - upload_started
            - committed
            - packet_generated
            - certificate_generated
            - snapshot_generated
            - share_created
            - share_viewed
            - dismiss_reminder
        createdAt:
          type: string
          format: date-time
        metadata:
          type: object
          description: Action-specific context (varies by action type)
          additionalProperties: true
    AuditLog:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - invalid_request
            - unauthorized
            - not_found
            - file_too_large
            - unsupported_file_type
            - version_not_committed
            - internal_error

  # ── Security ────────────────────────────────
  securitySchemes:
    x402:
      type: apiKey
      in: header
      name: X-PAYMENT-RESPONSE
      description: |
        x402 payment proof. Handled automatically by the Mirra
        platform — resource developers do not need to manage this.
    proofmarkApiKey:
      type: apiKey
      in: header
      name: X-ProofMark-Key
      description: |
        ProofMark API key. Auto-provisioned when a Mirra user
        first enables the ProofMark resource.
security:
  - x402: []
  - proofmarkApiKey: []
...